fix(suite): remove circular dependency in TroubleshootingTips
- Extracted TroubleshootingTipsItem and TroubleshootingTipsBaseProps types to separate types.ts file
- TroubleshootingTipsList now imports types directly instead of from TroubleshootingTips
- This breaks the circular import between TroubleshootingTips and TroubleshootingTipsList
- Resolves: packages/suite/src/components/suite/troubleshooting/TroubleshootingTips.tsx > packages/suite/src/components/suite/troubleshooting/TroubleshootingTipsList.tsx
